requirements.txt
azure-functions
grpcio>=1.2.0
azure.cosmos